requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
pymongo==4.5.0
motor>=3.3.0
redis>=5.0.0
qdrant-client>=1.14.1
python-dotenv>=1.0.0
bcrypt>=4.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.0.0
authlib>=1.3.0
httpx>=0.27.0
itsdangerous>=2.1.2
jinja2>=3.1.2

# AI Model API packages
openai>=1.12.0
anthropic>=0.22.0
google-generativeai>=0.7.0
cohere>=5.0.0
boto3>=1.34.0
requests>=2.31.0

# Database drivers
mysql-connector-python>=8.1.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
cryptography>=41.0.0  # For secure storage of OAuth tokens

# Integration packages
httpx>=0.27.0  # For HTTP requests to external APIs

# Content loader dependencies
beautifulsoup4>=4.12.0  # For HTML parsing
feedparser>=6.0.0  # For RSS feed parsing
lxml>=4.9.0  # For XML processing
google-api-python-client>=2.70.0  # For Google Drive API
google-auth-httplib2>=0.1.0  # For Google auth
google-auth-oauthlib>=1.0.0  # For Google OAuth

# Document processing dependencies
PyPDF2>=3.0.0  # For PDF text extraction
python-docx>=0.8.11  # For DOCX processing
python-pptx>=0.6.21  # For PowerPoint processing
textract>=1.6.5  # For extracting text from various document formats
pytz>=2023.3  # For timezone handling

# File storage and sharing dependencies
boto3>=1.28.0  # For AWS S3 storage
azure-storage-blob>=12.19.0  # For Azure Blob storage (optional)
python-magic>=0.4.27  # For better MIME type detection
pillow>=10.1.0  # For image processing
tqdm>=4.66.0  # For progress bars

# Phase 4: Production-Ready Dependencies
# Error Recovery & Circuit Breakers
aiohttp==3.9.1

# Monitoring & Metrics
psutil==5.9.6

# Performance Optimization
aioredis==2.0.1
motor==3.3.2
numpy==1.25.2

# Security Hardening
cryptography==41.0.7
PyJWT==2.8.0
bleach==6.1.0
email-validator==2.1.0